requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncio==3.4.3
redis==5.0.1
pydantic==2.5.0

# Consensus & Distributed Systems
loguru==0.7.2
dataclasses-json==0.6.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0

# Monitoring & Metrics
prometheus-client==0.19.0

# Docker & Deployment
python-dotenv==1.0.0

# Utilities
click==8.1.7
tabulate==0.9.0
psutil==5.9.6
